% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{area_pct_in_buffer}
\alias{area_pct_in_buffer}
\title{Calculate percentage of area type that are within certain buffer of stops}
\usage{
area_pct_in_buffer(
  stops,
  input_layer,
  proj,
  buffer_length,
  column,
  type,
  NA_omit = FALSE
)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polygon layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}

\item{column}{the character name of the column that you want to calculate in input_layer}

\item{type}{the character name of the type in the column that you want to calculate in input_layer}

\item{NA_omit}{TRUE or FALSE; whether omit NA value in the column that you want to calculate}
}
\value{
a polygon layer based on stops buffer with extra columns contain the results
}
\description{
Calculate percentage of area type that are within certain buffer of stops
}
\author{
Zehui Yin, \email{zehuiyin@gmail.com}
}
