% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_source_markers.R
\name{as_source_markers}
\alias{as_source_markers}
\alias{as_source_markers.sifkw_result}
\alias{as_source_markers.sif_result}
\alias{as_source_markers.find_files_result}
\title{Convert sifr result objects to RStudio source markers}
\usage{
as_source_markers(x)

\method{as_source_markers}{sifkw_result}(x)

\method{as_source_markers}{sif_result}(x)

\method{as_source_markers}{find_files_result}(x)
}
\arguments{
\item{x}{Any supported \R object}
}
\value{
An \link[rstudioapi:sourceMarkers]{rstudioapi::sourceMarkers} object
}
\description{
Convert sifr result objects to RStudio source markers
}
