% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{average_value_in_buffer}
\alias{average_value_in_buffer}
\title{Calculate average demographic value that are within certain buffer of stops}
\usage{
average_value_in_buffer(
  stops,
  input_layer,
  proj,
  buffer_length,
  column,
  NA_omit = FALSE
)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polygon layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}

\item{column}{the character name of the column that you want to calculate in input_layer}

\item{NA_omit}{TRUE or FALSE; whether omit NA value in the column that you want to calculate}
}
\value{
a point layer based on stops with one extra column contains the results
}
\description{
Calculate average demographic value that are within certain buffer of stops
}
\examples{
...
}
