% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate land use entropy for a certain buffer area}
\usage{
calculate_entropy(
  stops,
  input_layer,
  proj,
  buffer_length,
  column,
  exclude_intermediate = TRUE
)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polygon layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}

\item{column}{character column name that contains the land use categories}

\item{exclude_intermediate}{whether to exclude intermediate land use area percentage results in the output}
}
\value{
a point layer based on stops containing the entropy value for each stop
}
\description{
Calculate land use entropy for a certain buffer area
}
\details{
The formula used in this function to calculate entropy value is
\eqn{ Entropy = -\sum_{k=1}^nP_k\times\frac{ln(P_k)}{ln(n)}
}{ASCII}
}
\author{
Zehui Yin, \email{zehuiyin@gmail.com}
}
