% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{Find files}
\usage{
find_files(
  pattern,
  dir = ".",
  markers = interactive() && getOption("sifr.use_markers") &&
    requireNamespace("rstudioapi", quietly = TRUE),
  fixed = FALSE,
  case_sensitive = TRUE,
  path_pattern = getOption("sifr.path_pattern"),
  path_case_sensitive = FALSE,
  recursive = TRUE,
  encoding = "unknown"
)
}
\arguments{
\item{pattern}{\verb{character scalar}. A pattern for which to look in filenames}

\item{dir}{\code{character} scalar. A file system path to the directory in which
files should be search.}

\item{markers}{\code{logical} scalar. Display results in the RStudio source
markers pane.}

\item{fixed}{\code{logical} scalar. If \code{FALSE} \code{pattern} is
interpreted as regular expression, else the string is matched as-is.}

\item{case_sensitive}{\code{logical} scalar. If \code{TRUE} \code{pattern} is matched
case sensitively.}

\item{path_pattern}{\code{character} scalar. A regular expression pattern to match
file paths against. Defaults to common source files for R
(\code{.R}, \code{.Rmd}, \code{Rnw}, \code{Rhtml}, \code{c}, \code{cpp}). Please note that \code{.Rd} files
are \emph{not} included by default. You can modify the default behavior by
setting the \code{\link[=option]{option()}} \code{sifr.path_pattern}.}

\item{path_case_sensitive}{\code{logical}. If \code{TRUE} \code{path_pattern} is matched
case sensitively.}

\item{recursive}{\code{logical} scalar. If \code{TRUE} files are searched recursively
starting from \code{dir}.}

\item{encoding}{passed on to \code{\link[base:readLines]{base::readLines()}}}
}
\description{
Find files based on a regex pattern matched against the file path. Mainly
useful for interactive use.
}
\seealso{
\code{\link[=open_files]{open_files()}}
}
