% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{length_in_buffer}
\alias{length_in_buffer}
\title{Calculate length of input layer that are within certain buffer of stops}
\usage{
length_in_buffer(stops, input_layer, proj, buffer_length)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polyline layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}
}
\value{
a polygon layer based on stops buffer with extra columns contain the results
}
\description{
Calculate length of input layer that are within certain buffer of stops
}
\author{
Zehui Yin, \email{zehuiyin@gmail.com}
}
