% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{nearest_median_value}
\alias{nearest_median_value}
\title{Calculate median demographic value that are nearest to stops}
\usage{
nearest_median_value(stops, input_layer, column, proj)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polygon layer}

\item{column}{the character name of the column that you want to calculate in input_layer}

\item{proj}{numeric ESPG number of projection}
}
\value{
a point layer based on stops with one extra column contains the results
}
\description{
Calculate median demographic value that are nearest to stops
}
\examples{
...
}
