% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{open_files}
\alias{open_files}
\title{Open files}
\usage{
open_files(...)
}
\arguments{
\item{...}{either
\itemize{
\item \code{integer} vectors: open file(s) number x of the last \code{\link[=find_files]{find_files()}} result
\item \code{character} vectors: file paths
}}
}
\description{
Conveniently open files found via \code{\link[=find_files]{find_files()}}. Requires \pkg{rstudioapi}.
}
\examples{
\dontrun{

find_files("sif", markers = FALSE)

#> Results for sif
#>
#> 1 ./R/sif.R
#> 2 ./R/sif_results.R
#> 3 ./R/sifr-package.R
#> 4 ./tests/testthat/test_sif.R
#> 5 ./tests/testthat/testdata/sif_test.r
#> 6 ./tests/testthat/testdata/sif_test.rMd

open_files(3, 4)
}
}
\seealso{
\code{\link[=find_files]{find_files()}}
}
