% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files_result.R
\name{print.find_files_result}
\alias{print.find_files_result}
\title{Printing sif results}
\usage{
\method{print}{find_files_result}(x, markers = FALSE, ...)
}
\arguments{
\item{x}{a \code{find_files_result}}

\item{markers}{\code{logical} scalar. Display results in the RStudio source
markers pane.}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Printing sif results
}
