% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sif_results.R
\name{print.sif_result}
\alias{print.sif_result}
\alias{print.sifkw_result}
\title{Printing sif results}
\usage{
\method{print}{sif_result}(x, markers = FALSE, ...)

\method{print}{sifkw_result}(x, markers = FALSE, ...)
}
\arguments{
\item{x}{a \code{sif_result}}

\item{markers}{\code{logical} scalar. Display results in the RStudio source
markers pane.}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Printing sif results
}
