% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{return_input_layer_rows}
\alias{return_input_layer_rows}
\title{Retrieve rows from input layer with what_within_each_stops result}
\usage{
return_input_layer_rows(row, column, inter_stops, input_layer)
}
\arguments{
\item{row}{a single numeric value of the stop row}

\item{column}{the character of the column name that stores the intersect result; if using the result from the previous function, then it should be "input_rows_within"}

\item{inter_stops}{the name of the sf object and dataframe that store the intersect result}

\item{input_layer}{an polygon layer which you intersect with stops layer}
}
\value{
rows from input_layer based on the numeric row number just generated
}
\description{
Retrieve rows from input layer with what_within_each_stops result
}
\examples{
...
}
