% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_replace.R
\name{sed_replace}
\alias{sed_replace}
\title{Replace a matched regular expression pattern accross multiple files}
\usage{
sed_replace(pattern, replace, dir = ".", case_sensitive = TRUE,
  path_pattern = getOption("sifr.path_pattern"),
  path_case_sensitive = FALSE, recursive = TRUE)
}
\arguments{
\item{pattern}{\code{character} scalar. A \emph{sed} compatible regular expression
pattern for which to search in files.}

\item{replace}{\code{character} scalar.}

\item{dir}{\code{character} scalar. A file system path to the directory in which
files should be search.}

\item{case_sensitive}{\code{logical} scalar. If \code{TRUE} \code{pattern} is matched
case sensitvely. Please not that this is not supported on BSD
platforms such as macOS. See
\href{https://stackoverflow.com/questions/4412945/case-insensitive-search-and-replace-with-sed}{this stackoverflow post}.}

\item{path_pattern}{\code{character} scalar. A regular expression pattern to match
file paths against. Defaults to common source files for R
(\code{.R}, \code{.Rmd}, \code{Rnw}, \code{Rhtml}, \code{c}, \code{cpp}). Please note that \code{.Rd} files
are \emph{not} included by default. You can modify the default behaviour by
setting the \code{\link[=option]{option()}} \code{sifr.path_pattern}.}

\item{path_case_sensitive}{\code{logical}. If \code{TRUE} \code{path_pattern} is matched
case sensitively.}

\item{recursive}{\code{logical} scalar. If \code{TRUE} files are searched recursively
starting from \code{dir}.}
}
\value{
A \code{sif_result} \code{data.table} that contains all matched lines (before replacement took place)
}
\description{
Search and replace a regex pattern across several files. This only works on
POSIX compatible systems where \href{https://en.wikipedia.org/wiki/Sed}{sed}
is available. It is recommended to use this feature only on files checked
into a version control system like \emph{git}.
}
\section{Notes}{


\emph{sed} has its own flavour of regular expression that does not support all
features. The exact features supported are platform dependent (e.g.
case insensitive matching is not supported on macOS).
}

\section{Side Effects}{
 Replaces text that matches \code{pattern} accross all
files that match \code{path_pattern} with \code{replace}.
}

