% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{str_to_num}
\alias{str_to_num}
\title{Convert "input_rows_within" from string to numeric}
\usage{
str_to_num(row, column, dataframe)
}
\arguments{
\item{row}{a single numeric value of the stop row}

\item{column}{the character of the column name that stores the intersect result; if using the result from the previous function, then it should be "input_rows_within"}

\item{dataframe}{the name of the dataframe that store the result}
}
\value{
the numeric row number
}
\description{
Convert "input_rows_within" from string to numeric
}
\author{
Zehui Yin, \email{zehuiyin@gmail.com}
}
