% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{total_volume_in_buffer}
\alias{total_volume_in_buffer}
\title{Calculate total traffic volume within stop buffer areas}
\usage{
total_volume_in_buffer(stops, input_layer, proj, buffer_length, column)
}
\arguments{
\item{stops}{a point layer}

\item{input_layer}{an polyline layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}

\item{column}{character column name that contains traffic volume}
}
\value{
a polygon layer based on stops buffer with extra columns contain the results
}
\description{
Calculate total traffic volume within stop buffer areas
}
\examples{
...
}
