% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation_functions.R
\name{what_within_each_stops}
\alias{what_within_each_stops}
\title{Calculate what rows of input layer fall within every stops buffer area}
\usage{
what_within_each_stops(stops, input_layer, proj, buffer_length)
}
\arguments{
\item{stops}{an point layer}

\item{input_layer}{an polygon layer}

\item{proj}{numeric ESPG number of projection}

\item{buffer_length}{an single numeric value (note: the unit is based purely on the projection)}
}
\value{
a point layer based on stops with one extra column contains the results
}
\description{
Calculate what rows of input layer fall within every stops buffer area
}
\examples{
...
}
