% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifr-package.R
\docType{package}
\name{sifr-package}
\alias{sifr}
\alias{sifr-package}
\title{sifr: Search and Replace in R Source Files}
\description{
Search and replace regex patterns across R source files (and other files).
}
\section{Options}{


\describe{
\item{\code{sifr.path_pattern}}{A \code{character} scalar containing a \code{regex} pattern
to match file paths against. This is used as the default for \code{\link[=sif]{sif()}}
and co.}
}
}

\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
